Updated: organize page.tsx
